+++ apps/client/src/app/pages/portfolio/activities/activities-page.routes.ts
@@ -1,12 +1,11 @@
 import { AuthGuard } from '@ghostfolio/client/core/auth.guard';
 import { internalRoutes } from '@ghostfolio/common/routes/routes';
 
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
+import { Routes } from '@angular/router';
 
 import { ActivitiesPageComponent } from './activities-page.component';
 
-const routes: Routes = [
+export const routes: Routes = [
   {
     canActivate: [AuthGuard],
     component: ActivitiesPageComponent,
@@ -14,9 +13,3 @@ const routes: Routes = [
     title: internalRoutes.portfolio.subRoutes.activities.title
   }
 ];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class ActivitiesPageRoutingModule {}